backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
anthropic==0.28.0
pdfplumber==0.11.1
python-multipart==0.0.9
pydantic==2.7.4
python-dotenv==1.0.1
